implment namespace for manager and admin
