+++ app/controllers/admin/base_controller.rb
@@ -3,7 +3,7 @@ module Admin
     before_action :validate_admin
     
     def validate_admin
-      return redirect_to root_path unless current_user
+      return redirect_to root_path unless current_user.admin?
     end
   end
 end
