+++ app/views/admin/users/_form.html.erb
@@ -1,6 +1,6 @@
 <div class="container float-right">                                    
   <%= form_for [:admin, @user] do |f| %>
-	<div class="container">
+  <div class="container">
     <% if @user.errors.any? %>
       <% @user.errors.full_messages.each  do |error| %>
         <li> <%= error %></li>
@@ -8,37 +8,37 @@
     <% end %>
   </div>
     <div class="row">
-			<div class="col-md-10 form-group ">
-					<%= f.text_field :first_name,  class:"form-control", id:"form_uname", placeholder:"First Name *", required: true %>  
-					<span class="error_form" id="uname_error_message"></span>
-			</div>
+      <div class="col-md-10 form-group ">
+          <%= f.text_field :first_name,  class:"form-control", id:"form_uname", placeholder:"First Name *", required: true %>  
+          <span class="error_form" id="uname_error_message"></span>
+      </div>
 
-			<div class="col-md-10 form-group">
-					<%= f.text_field :last_name,  class:"form-control", id:"form_lname" ,placeholder:" Last Name *", required: true %>
-					<span class="error_form" id="lname_error_message"></span>
-			</div> 
+      <div class="col-md-10 form-group">
+          <%= f.text_field :last_name,  class:"form-control", id:"form_lname" ,placeholder:" Last Name *", required: true %>
+          <span class="error_form" id="lname_error_message"></span>
+      </div> 
 
-			<div class="col-md-10 form-group">
-					<%= f.email_field :email, class:"form-control", id: "form_email", placeholder:"Your Eamil *", required: true %> 
-					<span class="error_form" id="email_error_message"></span>  
-			</div>
+      <div class="col-md-10 form-group">
+          <%= f.email_field :email, class:"form-control", id: "form_email", placeholder:"Your Eamil *", required: true %> 
+          <span class="error_form" id="email_error_message"></span>  
+      </div>
 
-			<div class="col-md-10 form-group">
-					<%= f.password_field :password, class:"form-control", id:"form_password", placeholder:"Password *", required: true %>
-					<span class="error_form" id="password_error_message"></span> 
-			</div>
+      <div class="col-md-10 form-group">
+          <%= f.password_field :password, class:"form-control", id:"form_password", placeholder:"Password *", required: true %>
+          <span class="error_form" id="password_error_message"></span> 
+      </div>
 
-			<div class="col-md-10 form-group">
-					<%= f.password_field :password_confirmation ,class:"form-control", id: "form_retry_password", placeholder:"Retry Password *", required: true %> 
-					<span class="error_form" id="retry_password_error_message"></span>      
-			</div> 
-			
-			<div class="col-md-10 form-group">
-				<%= f.file_field :profile_image %>
-				<span class="error_form" id="password_error_message"></span> 
-		  </div>
-		 
-		</div>     
-		<%= f.submit  class:"btn btn-secondary" %>	                						
-	<% end %>
+      <div class="col-md-10 form-group">
+          <%= f.password_field :password_confirmation ,class:"form-control", id: "form_retry_password", placeholder:"Retry Password *", required: true %> 
+          <span class="error_form" id="retry_password_error_message"></span>      
+      </div> 
+      
+      <div class="col-md-10 form-group">
+        <%= f.file_field :profile_image %>
+        <span class="error_form" id="password_error_message"></span> 
+      </div>
+    
+    </div>     
+    <%= f.submit  class:"btn btn-secondary" %>	                						
+  <% end %>
 </div>
