+++ app/views/manager/clients/_client.json.jbuilder
