+++ app/views/manager/clients/_form.html.erb
@@ -1,5 +1,5 @@
 <div class="container">
-  <%= form_with(model: client) do |form| %>
+  <%= form_for([:manager, client]) do |form| %>
     <% if client.errors.any? %>
       <div id="error_explanation">
         <h2><%= pluralize(client.errors.count, "error") %> prohibited this client from being saved:</h2>
@@ -11,30 +11,30 @@
         </ul>
       </div>
     <% end %>
-    <%= link_to 'Back', clients_path, class: "btn btn-success float-right"%><br>
+    <%= link_to 'Back', manager_clients_path, class: "btn btn-success float-right"%><br>
     <div class="form-group">
       <%= form.label :first_name %>
-      <%= form.text_field :first_name , class: "form-control"%>
+      <%= form.text_field :first_name , class: "form-control", required: true %>
     </div>
 
     <div class="form-group">
       <%= form.label :last_name %>
-      <%= form.text_field :last_name , class: "form-control"%>
+      <%= form.text_field :last_name , class: "form-control", required: true %>
     </div>
 
     <div class="form-group">
       <%= form.label :address %>
-      <%= form.text_area :address , class: "form-control"%>
+      <%= form.text_area :address , class: "form-control", required: true %>
     </div>
 
     <div class="form-group">
       <%= form.label :phone_number %>
-      <%= form.text_field :phone_number , class: "form-control"%>
+      <%= form.text_field :phone_number , class: "form-control", required: true %>
     </div>
 
     <div class="form-group">
       <%= form.label :email %>
-      <%= form.text_field :email , class: "form-control"%>
+      <%= form.email_field :email , class: "form-control", required: true %>
     </div>
 
     <div class="actions">
