+++ app/views/manager/clients/edit.html.erb
@@ -1,0 +1,6 @@
+<h1>Editing Client</h1>
+
+<%= render 'form', client: @client %>
+
+<%= link_to 'Show', manager_client_path(@client) %> |
+<%= link_to 'Back', manager_clients_path %>
