+++ app/views/manager/clients/index.json.jbuilder
