+++ app/views/manager/clients/show.html.erb
@@ -25,5 +25,5 @@
   <%= @client.email %>
 </p>
 
-<%= link_to 'Edit', edit_client_path(@client) %> |
-<%= link_to 'Back', clients_path %>
+<%= link_to 'Edit', edit_manager_client_path(@client) %> |
+<%= link_to 'Back', manager_clients_path %>
