+++ app/views/manager/clients/show.json.jbuilder
