+++ app/views/manager/comments/_comment.html.erb
