+++ app/views/manager/comments/_form.html.erb
@@ -1,4 +1,4 @@
-<%= form_for([:admin, @project, @project.comments.build], remote: true) do |f| %>
+<%= form_for([:manager, @project, @project.comments.build], remote: true) do |f| %>
   <h1><%= f.label :comment %></h1><br />
   <%= f.text_area :content %><br />
   <%= f.submit %>
