+++ app/views/manager/comments/create.js.erb
@@ -1,0 +1,1 @@
+$("#comments-section").append("<%= escape_javascript(render 'manager/comments/comment', comment: @comment) %>")
