+++ app/views/manager/projects/index.html.erb
@@ -14,10 +14,10 @@
           <td><%= project.name %></td>
           <td><%= project.description %></td>
           <td><%= project.created_at%></td>
-          <td><%= link_to 'Show', admin_project_path(project), class:"btn btn-light" %></td>
+          <td><%= link_to 'Show', manager_project_path(project), class:"btn btn-light" %></td>
           <% if user_signed_in? %>
-          <td> <%= link_to 'Edit', edit_admin_project_path(project), class:"btn btn-light" %></td>
-          <td><%= link_to 'Delete', admin_project_path(project), method: :delete, data: { confirm: 'Are you sure?' } , class:"btn btn-danger"%></td>
+          <td> <%= link_to 'Edit', edit_manager_project_path(project), class:"btn btn-light" %></td>
+          <td><%= link_to 'Delete', manager_project_path(project), method: :delete, data: { confirm: 'Are you sure?' } , class:"btn btn-danger"%></td>
           <% end %>
         </tr>
       <% end %>
