+++ db/migrate/20210625071444_add_is_active_to_users.rb
@@ -1,0 +1,5 @@
+class AddIsActiveToUsers < ActiveRecord::Migration[6.1]
+  def change
+    add_column :users, :is_active, :boolean, default: false, null: false
+  end
+end
